separated word list and lists
